requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncer==0.0.17
attrs==26.1.0
beautifulsoup4==4.14.3
bidict==0.23.1
cachetools==6.2.6
certifi==2026.4.22
cffi==2.0.0
chainlit==2.11.1
chardet==7.4.3
charset-normalizer==3.4.7
chevron==0.14.0
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
cuid==0.4
dataclasses-json==0.6.7
ddgs==9.14.1
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.136.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.73.1
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.5.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.13
importlib_metadata==8.7.1
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-google-genai==4.2.2
langchain-protocol==0.0.14
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.12
langgraph-sdk==0.3.13
langsmith==0.7.38
Lazify==0.4.0
literalai==0.1.201
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.4
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-exporter-otlp-proto-http==1.41.1
opentelemetry-instrumentation==0.62b1
opentelemetry-instrumentation-agno==0.60.0
opentelemetry-instrumentation-alephalpha==0.60.0
opentelemetry-instrumentation-anthropic==0.60.0
opentelemetry-instrumentation-bedrock==0.60.0
opentelemetry-instrumentation-chromadb==0.60.0
opentelemetry-instrumentation-cohere==0.60.0
opentelemetry-instrumentation-crewai==0.60.0
opentelemetry-instrumentation-google-generativeai==0.60.0
opentelemetry-instrumentation-groq==0.60.0
opentelemetry-instrumentation-haystack==0.60.0
opentelemetry-instrumentation-lancedb==0.60.0
opentelemetry-instrumentation-langchain==0.60.0
opentelemetry-instrumentation-llamaindex==0.60.0
opentelemetry-instrumentation-logging==0.62b1
opentelemetry-instrumentation-marqo==0.60.0
opentelemetry-instrumentation-mcp==0.60.0
opentelemetry-instrumentation-milvus==0.60.0
opentelemetry-instrumentation-mistralai==0.60.0
opentelemetry-instrumentation-ollama==0.60.0
opentelemetry-instrumentation-openai==0.60.0
opentelemetry-instrumentation-openai-agents==0.60.0
opentelemetry-instrumentation-pinecone==0.60.0
opentelemetry-instrumentation-qdrant==0.60.0
opentelemetry-instrumentation-redis==0.62b1
opentelemetry-instrumentation-replicate==0.60.0
opentelemetry-instrumentation-requests==0.62b1
opentelemetry-instrumentation-sagemaker==0.60.0
opentelemetry-instrumentation-sqlalchemy==0.62b1
opentelemetry-instrumentation-threading==0.62b1
opentelemetry-instrumentation-together==0.60.0
opentelemetry-instrumentation-transformers==0.60.0
opentelemetry-instrumentation-urllib3==0.62b1
opentelemetry-instrumentation-vertexai==0.60.0
opentelemetry-instrumentation-voyageai==0.60.0
opentelemetry-instrumentation-watsonx==0.60.0
opentelemetry-instrumentation-weaviate==0.60.0
opentelemetry-instrumentation-writer==0.60.0
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
opentelemetry-semantic-conventions-ai==0.5.1
opentelemetry-util-http==0.62b1
orjson==3.11.8
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==26.2
pandas==3.0.2
pillow==12.2.0
playwright==1.59.0
postgrest==2.29.0
primp==1.2.3
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyee==13.0.1
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.27
python-socketio==5.16.1
PyYAML==6.0.3
realtime==2.29.0
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
selenium==4.43.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
sse-starlette==3.4.1
starlette==1.0.0
storage3==2.29.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.29.0
supabase-auth==2.29.0
supabase-functions==2.29.0
syncer==2.0.3
tenacity==9.1.4
threadpoolctl==3.6.0
tomli==2.4.1
tqdm==4.67.3
traceloop-sdk==0.60.0
trio==0.33.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.2
wsproto==1.3.2
xgboost==3.2.0
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
